backend/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
fuzzywuzzy==0.18.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Levenshtein==0.26.1
MarkupSafe==3.0.2
numpy==2.1.3
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.1
python-whois==0.9.5
RapidFuzz==3.10.1
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
threadpoolctl==3.5.0
urllib3==2.2.3
watchdog==6.0.0
Werkzeug==3.1.2
